requirements.txt
streamlit
bcrypt
pyotp
qrcode[pil]
pillow
